requirements.txt
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.115.4
pydantic==2.9.2
uvicorn==0.32.0
selenium==4.11.2